2023/03/17納品分（フェーズ２） ■送付物 ・ソースコード 　TMQ6WebSolution：TMQ本体のソースコードです 　TMQSendMail：長期計画メール送信機能のソースコードです 　TMQSendMailHistory：変更管理メール送信機能のソースコードです ・実行環境 　TMQ2022：IIS公開フォルダに配置する実行環境です 　長期計画メール送信：長期計画メール送信機能の実行ファイルです 　変更管理メール送信：変更管理メール送信機能の実行ファイルです 　TMQ_NEW.bak：DBのバックアップファイルです。以降はこちらのDBをベースとしてデータ投入をお願い致します。 ・確認チェックリスト 　上記の実行環境にて確認を行ったチェックリストです
上記とは別に、Teamsの所定のフォルダのドキュメントを更新しています
・基本設計書
　ファイルやりとり専用>基本設計書
・詳細設計書
　ファイルやりとり専用>詳細設計書
・単体テスト仕様書
　ファイルやりとり専用>単体テスト仕様書
　ExcelPortフォルダと変更管理フォルダ
・単体テスト課題管理表
　ファイルやりとり専用>課題管理表>ATTSフェーズ2単体テスト課題管理表.xlsx
・内部結合テスト仕様書
　ファイルやりとり専用>フェーズ2内部結合テスト
・内部結合テスト課題管理表
　ファイルやりとり専用>課題管理表>ATTSフェーズ2内部結合テスト課題管理表.xlsx
・マニュアル
　ファイルやり取り専用>マニュアル

■ExcelPortテンプレートブックのパスワードについて
テンプレートブックは「template_RP1000.xlsm」となります。
ブック保護およびソースのパスワードロックのパスワードは暫定的に「aec」としております。
また、VBAソース内にもパスワードを固定で保持しております。AEC様にてブック保護のパスワードを変更する際はこちらも併せて変更ください。(modCommonモジュールの218行目です。)

+++ TMQ6WebSolution/BackEndSolution/BusinessLogic_DM0002/BusinessLogic_DM0002.cs
@@ -491,7 +491,7 @@ namespace BusinessLogic_DM0002
         /// <param name="keyId">キーID</param>
         /// <param name="documentTypeValNo">表示している文書種類コンボボックスの項目番号(VAL)</param>
         /// <returns>エラーの場合False</returns>
-        private bool getSubject(string sqlName, int functionTypeId, int keyId, int documentTypeValNo)
+        private bool getSubject(string sqlName, int functionTypeId, long keyId, int documentTypeValNo)
         {
             // 件名取得SQL文の取得
             TMQUtil.GetFixedSqlStatement(SqlName.SubDir, sqlName, out string outSql);
